servers changed

+++ app/models/Servers.php
@@ -50,8 +50,14 @@ class Servers extends BaseModel
             return false;
         }
         $password = preg_replace('/[\x00]+/', '', $this->getDI()->getCrypt()->decryptBase64($this->password));
-        $out = `sshpass -p "{$password}" ssh {$this->username}@{$this->hostname} -p {$this->port} "mkdir ~/{$dir} && chmod 0755 ~/{$dir}"`;
-        return ($out ? false : true);
+        if (file_exists(APP_DIR . '/cache/servers/' . $this->username . '/mount')) {
+            mkdir(APP_DIR . '/cache/servers/' . $this->username . '/' . $dir, 0755);
+            chmod(APP_DIR . '/cache/servers/' . $this->username . '/' . $dir, 0755);
+            return (file_exists(APP_DIR . '/cache/servers/' . $this->username . '/' . $dir) ? true : false);
+        } else {
+            $out = `sshpass -p "{$password}" ssh {$this->username}@{$this->hostname} -p {$this->port} "mkdir ~/{$dir} && chmod 0755 ~/{$dir}"`;
+            return ($out ? false : true);
+        }
     }
 
     public function rmdir($dir = '')
@@ -59,9 +65,14 @@ class Servers extends BaseModel
         if (!is_numeric($dir)) {
             return false;
         }
-        $password = preg_replace('/[\x00]+/', '', $this->getDI()->getCrypt()->decryptBase64($this->password));
-        $out = `sshpass -p "{$password}" ssh {$this->username}@{$this->hostname} -p {$this->port} "rm -rf ~/{$dir}"`;
-        return ($out ? false : true);
+        if (file_exists(APP_DIR . '/cache/servers/' . $this->username . '/mount')) {
+            self::delTree(APP_DIR . '/cache/servers/' . $this->username . '/' . $dir);
+            return (file_exists(APP_DIR . '/cache/servers/' . $this->username . '/' . $dir) ? false : true);
+        } else {
+            $password = preg_replace('/[\x00]+/', '', $this->getDI()->getCrypt()->decryptBase64($this->password));
+            $out = `sshpass -p "{$password}" ssh {$this->username}@{$this->hostname} -p {$this->port} "rm -rf ~/{$dir}"`;
+            return ($out ? false : true);
+        }
     }
 
     public function ls($dir = '')
@@ -70,14 +81,23 @@ class Servers extends BaseModel
         if ($dir && !is_numeric($dir)) {
             return $result;
         }
-        $password = preg_replace('/[\x00]+/', '', $this->getDI()->getCrypt()->decryptBase64($this->password));
-        $command = "sshpass -p '{$password}' ssh {$this->username}@{$this->hostname} -p {$this->port} 'ls ~/{$dir}'";
-        $result = array();
-        exec($command, $result);
-        if (!is_array($result)) {
+        if (file_exists(APP_DIR . '/cache/servers/' . $this->username . '/mount')) {
+            $result = array_diff(scandir(APP_DIR . '/cache/servers/' . $this->username . '/' . $dir), array('.','..'));
+            if (!is_array($result)) {
+                $result = array();
+            }
+            return $result;
+        } else {
+            $password = preg_replace('/[\x00]+/', '', $this->getDI()->getCrypt()->decryptBase64($this->password));
+            $command = "sshpass -p '{$password}' ssh {$this->username}@{$this->hostname} -p {$this->port} 'ls ~/{$dir}'";
             $result = array();
+            exec($command, $result);
+            if (!is_array($result)) {
+                $result = array();
+            }
+            return $result;
         }
-        return $result;
+        
     }
 
     public function du($dir = '')
@@ -85,8 +105,14 @@ class Servers extends BaseModel
         if ($dir && !is_numeric($dir)) {
             return false;
         }
-        $password = preg_replace('/[\x00]+/', '', $this->getDI()->getCrypt()->decryptBase64($this->password));
-        $command = "sshpass -p '{$password}' ssh {$this->username}@{$this->hostname} -p {$this->port} 'du -s ~/{$dir}'";
+        if (file_exists(APP_DIR . '/cache/servers/' . $this->username . '/mount')) {
+            $dir = APP_DIR . '/cache/servers/' . $this->username . '/' . $dir;
+            $command = "du -s {$dir}";
+        } else {
+            $password = preg_replace('/[\x00]+/', '', $this->getDI()->getCrypt()->decryptBase64($this->password));
+            $command = "sshpass -p '{$password}' ssh {$this->username}@{$this->hostname} -p {$this->port} 'du -s ~/{$dir}'";
+        }
+        
         $out = array();
         exec($command, $out);
         if (!is_array($out)) {
@@ -101,8 +127,13 @@ class Servers extends BaseModel
         if ($dir && !is_numeric($dir)) {
             return $result;
         }
-        $password = preg_replace('/[\x00]+/', '', $this->getDI()->getCrypt()->decryptBase64($this->password));
-        $command = "sshpass -p '{$password}' rsync -ptrzv -e 'ssh -p {$this->port}' --remove-source-files --progress {$source} {$this->username}@{$this->hostname}:~/{$dir}";
+        if (file_exists(APP_DIR . '/cache/servers/' . $this->username . '/mount')) {
+            $dir = APP_DIR . '/cache/servers/' . $this->username . '/' . $dir;
+            $command = "rsync -ptrzv --remove-source-files --progress {$source} {$dir}";
+        } else {
+            $password = preg_replace('/[\x00]+/', '', $this->getDI()->getCrypt()->decryptBase64($this->password));
+            $command = "sshpass -p '{$password}' rsync -ptrzv -e 'ssh -p {$this->port}' --remove-source-files --progress {$source} {$this->username}@{$this->hostname}:~/{$dir}";
+        }
         $out = array();
         exec($command, $out);
         if (is_array($out)) {
@@ -114,4 +145,11 @@ class Servers extends BaseModel
         }
         return $result;
     }
+    public static function delTree($dir) {
+        $files = array_diff(scandir($dir), array('.','..'));
+        foreach ($files as $file) {
+            (is_dir("$dir/$file")) ? self::delTree("$dir/$file") : unlink("$dir/$file");
+        }
+        return rmdir($dir);
+    } 
 }
